pyproject.toml
[tool.poetry]
name = "barbiecam"
version = "0.1.0"
description = ""
authors = ["pajowu <git@ca.pajowu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
pillow = "^10.1.0"
opencv-python = "^4.8.1.78"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
